+++ src/codec/request/call.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, io::Write};
+use std::io::Write;
 
 use crate::{
     codec::{
@@ -12,12 +12,12 @@ use crate::{
 use super::Request;
 
 #[derive(Clone, Debug)]
-pub(crate) struct Call<T> {
-    pub function_name: Cow<'static, str>,
+pub(crate) struct Call<'a, T> {
+    pub function_name: &'a str,
     pub tuple: T,
 }
 
-impl<T: Tuple> Request for Call<T> {
+impl<'a, T: Tuple> Request for Call<'a, T> {
     fn request_type() -> RequestType
     where
         Self: Sized,
@@ -28,16 +28,16 @@ impl<T: Tuple> Request for Call<T> {
     // NOTE: `&mut buf: mut` is required since I don't get why compiler complain
     fn encode(&self, mut buf: &mut dyn Write) -> Result<(), EncodingError> {
         rmp::encode::write_map_len(&mut buf, 2)?;
-        write_kv_str(buf, keys::FUNCTION_NAME, self.function_name.as_ref())?;
+        write_kv_str(buf, keys::FUNCTION_NAME, self.function_name)?;
         write_kv_tuple(buf, keys::TUPLE, &self.tuple)?;
         Ok(())
     }
 }
 
-impl<T> Call<T> {
-    pub(crate) fn new(function_name: impl Into<Cow<'static, str>>, args: T) -> Self {
+impl<'a, T> Call<'a, T> {
+    pub(crate) fn new(function_name: &'a str, args: T) -> Self {
         Self {
-            function_name: function_name.into(),
+            function_name,
             tuple: args,
         }
     }