+++ src/codec/request/eval.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, io::Write};
+use std::io::Write;
 
 use crate::{
     codec::{
@@ -12,12 +12,12 @@ use crate::{
 use super::Request;
 
 #[derive(Clone, Debug)]
-pub(crate) struct Eval<T> {
-    pub expr: Cow<'static, str>,
+pub(crate) struct Eval<'a, T> {
+    pub expr: &'a str,
     pub tuple: T,
 }
 
-impl<T: Tuple> Request for Eval<T> {
+impl<'a, T: Tuple> Request for Eval<'a, T> {
     fn request_type() -> RequestType
     where
         Self: Sized,
@@ -28,18 +28,14 @@ impl<T: Tuple> Request for Eval<T> {
     // NOTE: `&mut buf: mut` is required since I don't get why compiler complain
     fn encode(&self, mut buf: &mut dyn Write) -> Result<(), EncodingError> {
         rmp::encode::write_map_len(&mut buf, 2)?;
-        write_kv_str(buf, keys::EXPR, self.expr.as_ref())?;
+        write_kv_str(buf, keys::EXPR, self.expr)?;
         write_kv_tuple(buf, keys::TUPLE, &self.tuple)?;
         Ok(())
     }
 }
 
-impl<T> Eval<T> {
-    // TODO: introduce some convenient way to pass arguments
-    pub fn new(expr: impl Into<Cow<'static, str>>, args: T) -> Self {
-        Self {
-            expr: expr.into(),
-            tuple: args,
-        }
+impl<'a, T> Eval<'a, T> {
+    pub fn new(expr: &'a str, args: T) -> Self {
+        Self { expr, tuple: args }
     }
 }